refactor(backend): migrate ClaudeWSController + ClaudeSessionManager to DI, remove shims (phase 3 step 6)
